feat: setting multi module
