+++ core/data/src/main/java/me/seokang/commerce/data/MyClass.kt
@@ -1,0 +1,4 @@
+package me.seokang.commerce.data
+
+class MyClass {
+}
