+++ core/domain/src/main/java/me/seokang/commerce/domain/MyClass.kt
@@ -1,0 +1,4 @@
+package me.seokang.commerce.domain
+
+class MyClass {
+}
